requirement.txt
langchain
langchain-core
langchain-google-genai
langgraph
streamlit
python-dotenv